fix: merge guardian and guardian_ninjas contexts
